refactor: rename src/__test__ to src/__tests__
