--- src/__test__/clustering.spec.ts
@@ -1,11 +1,0 @@
-import { onMessage, send } from '../clustering';
-
-describe('clustering test', () => {
-  test('send should be defined', () => {
-    expect(send).toBeInstanceOf(Function);
-  });
-
-  test('onMessage should be defined', () => {
-    expect(onMessage).toBeInstanceOf(Function);
-  });
-});
