+++ src/__tests__/categories.spec.ts
