+++ src/index.ts
@@ -1,5 +1,4 @@
-export { Category } from './categories';
-export { Configuration } from './configure/configure';
+export { Configuration } from './configure';
 export { LevelType } from './levels';
 
 export { getLogger, configure, logger } from './log4js';
